refactor: enforce gradebook workflow boundary
